requirements.txt
flask
flask-bcrypt
mysql-connector-python
python-dotenv
